Use "int" for user counters.
git-svn-id: svn://svn.code.sf.net/p/polyml/code/trunk/polyml@1577 ae7d391e-3f74-4a2b-a0db-a8776840fd2a

+++ libpolyml/statistics.cpp
@@ -328,7 +328,7 @@ void Statistics::updatePeriodicStats(POLYUNSIGNED freeWords, unsigned threadsInM
     }
 }
 
-void Statistics::setUserCounter(unsigned which, intptr_t value)
+void Statistics::setUserCounter(unsigned which, int value)
 {
     if (statMemory)
     {
